requirements.txt
pdf2image==1.16.3
Pillow==10.1.0
speech_recognition==3.10.1
unidecode==1.3.7